requirements.txt
asgiref==3.7.2
Django==3.2
django-cors-headers==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
Pillow==10.0.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
